Measurement time
Making sure that the code measures are correct. The problem is that the userscript is too fast for the animations to even run each time the mouse event is triggerd, so the elapsed time is written out as null but if the userscript is diabled and the mouse event is triggerd manually the elapsed time seems fine with the console time-function.